Early Testing/Lucentia v100-Bravo/Lucentia/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
plaid-python==37.1.0
pydantic==2.6.3
pydantic-core==2.16.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
asyncpg==0.29.0
python-dotenv==1.0.0